requirements.txt
fastapi>=0.101.0
uvicorn>=0.23.2
pydantic>=2.1.0
groq>=0.4.0
python-dotenv>=1.0.0
